refactor: импорты модулей для запуска с помощь баш команды (uvicorn src.main:app --host 0.0.0.0 --port 8050 --reload)
