+++ src/db/aiosqlite.py
@@ -2,7 +2,7 @@ from typing import AsyncGenerator
 
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from core.settings import settings
+from src.core.settings import settings
 
 # Создание асинхронного движка
 engine = create_async_engine(settings.database_url_async, echo=True)
