+++ src/main.py
@@ -4,8 +4,8 @@ from fastapi.encoders import jsonable_encoder
 from fastapi.exceptions import RequestValidationError
 from fastapi.responses import ORJSONResponse
 
-from api.v1 import bot_view, webhook_view
-from core.settings import settings
+from src.api.v1 import bot_view, webhook_view
+from src.core.settings import settings
 
 app = FastAPI(
     title=settings.PROJECT_NAME,
