+++ src/telegram_bot/bot_main.py
@@ -4,8 +4,8 @@ from aiogram import Bot, Dispatcher
 from aiogram.client.default import DefaultBotProperties
 from aiogram.enums import ParseMode
 
-from core.settings import settings
-from telegram_bot.routers import main_router
+from src.core.settings import settings
+from src.telegram_bot.routers import main_router
 
 # Создаем экземпляр бота с использованием токена из конфигурации
 bot = Bot(token=settings.BOT_TOKEN, default=DefaultBotProperties(parse_mode=ParseMode.HTML))
